Put protocol field definitions in a separate file
